Initial commit with first tries of Tasks 1 and 2
First commit for this project, in fulfillment of the City of Cape Town
Data Science Unit Code Challenge for Data Engineering appllicants.

I've made a first go at the Tasks 1 and 2 which are in
data_extraction.R and initial_data_transformation.R. They still need to
be optimised and cleaned up and made possible to run out of the box.
But first I will work on the final two tasks.

I've also updated the README.md and moved the Code Challenge
specifications to CHALLENGESPECS.md.