Added pdf generated reports from the R Markdown files
Added the pdf versions of the R Markdown generated reports for each
of the tasks. As well as a final report in pdf format with them
all merged together. Also fixed a few minor bugs:
* data_loading_task was attempting to upload to my personal bucket
as opposed to the challenge bucket, uncommented.
* num_lines in further_data_transformations_report.Rmd was set to
10000, shouuld be Inf to process the whole file
* included  eval=FALSE for the R code chunks in
initial_data_transformation_report.Rmd and data_loading_tasks.Rmd so
that they aren't run and cause the document to not generate. The
code to run for those two tasks is initial_data_transformation.R
and data_loading.R